Move Salesforce Mapping UI into its own module
