+++ modules/salesforce_mapping/salesforce_mapping.info.yml
@@ -3,7 +3,6 @@ type: module
 description: Map Drupal entities to Salesforce objects.
 package: Salesforce
 core: 8.x
-configure: entity.salesforce_mapping.list
 dependencies:
   - salesforce:salesforce
   - dynamic_entity_reference:dynamic_entity_reference (2.x)
