+++ modules/salesforce_mapping_ui/config/optional/views.view.salesforce_mapped_objects.yml
