+++ modules/salesforce_mapping_ui/salesforce_mapping_ui.info.yml
@@ -1,0 +1,8 @@
+name: Salesforce Mapping UI
+type: module
+description: User interface for managing Salesforce mappings.
+package: Salesforce
+core: 8.x
+configure: entity.salesforce_mapping.list
+dependencies:
+  - salesforce:salesforce_mapping
