+++ modules/salesforce_mapping_ui/salesforce_mapping_ui.links.contextual.yml
