+++ modules/salesforce_mapping_ui/src/Form/SalesforceMappingAddForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\salesforce_mapping\Form;
+namespace Drupal\salesforce_mapping_ui\Form;
 
 use Drupal\Core\Form\FormStateInterface;
