+++ modules/salesforce_mapping_ui/src/Form/SalesforceMappingDisableForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\salesforce_mapping\Form;
+namespace Drupal\salesforce_mapping_ui\Form;
 
 use Drupal\Core\Entity\EntityConfirmFormBase;
 use Drupal\Core\Form\FormStateInterface;
@@ -9,20 +9,20 @@ use Drupal\Core\Url;
 /**
  * Salesforce Mapping Disable Form .
  */
-class SalesforceMappingEnableForm extends EntityConfirmFormBase {
+class SalesforceMappingDisableForm extends EntityConfirmFormBase {
 
   /**
    * {@inheritdoc}
    */
   public function getQuestion() {
-    return $this->t('Are you sure you want to enable the mapping %name?', ['%name' => $this->entity->label()]);
+    return $this->t('Are you sure you want to disable the mapping %name?', ['%name' => $this->entity->label()]);
   }
 
   /**
    * {@inheritdoc}
    */
   public function getDescription() {
-    return $this->t('Enabling a mapping will restart any automatic synchronization.');
+    return $this->t('Disabling a mapping will stop any automatic synchronization and hide the mapping.');
   }
 
   /**
@@ -36,7 +36,7 @@ class SalesforceMappingEnableForm extends EntityConfirmFormBase {
    * {@inheritdoc}
    */
   public function getConfirmText() {
-    return $this->t('Enable');
+    return $this->t('Disable');
   }
 
   /**
@@ -45,10 +45,9 @@ class SalesforceMappingEnableForm extends EntityConfirmFormBase {
   public function submit(array $form, FormStateInterface $form_state) {
     parent::submit($form, $form_state);
 
-    $this->entity->enable()->save();
+    $this->entity->disable()->save();
     $form_state['redirect_route'] = [
-      'route_name' => 'entity.salesforce_mapping.edit_form',
-      'route_parameters' => ['salesforce_mapping' => $this->entity->id()],
+      'route_name' => 'entity.salesforce_mapping.list',
     ];
   }
