+++ modules/salesforce_mapping_ui/src/Form/SalesforceMappingEditForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\salesforce_mapping\Form;
+namespace Drupal\salesforce_mapping_ui\Form;
 
 /**
  * Salesforce Mapping Add/Edit Form.
