+++ modules/salesforce_mapping_ui/src/Form/SalesforceMappingEnableForm.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\salesforce_mapping\Form;
+namespace Drupal\salesforce_mapping_ui\Form;
 
 use Drupal\Core\Entity\EntityConfirmFormBase;
 use Drupal\Core\Form\FormStateInterface;
@@ -9,20 +9,20 @@ use Drupal\Core\Url;
 /**
  * Salesforce Mapping Disable Form .
  */
-class SalesforceMappingDisableForm extends EntityConfirmFormBase {
+class SalesforceMappingEnableForm extends EntityConfirmFormBase {
 
   /**
    * {@inheritdoc}
    */
   public function getQuestion() {
-    return $this->t('Are you sure you want to disable the mapping %name?', ['%name' => $this->entity->label()]);
+    return $this->t('Are you sure you want to enable the mapping %name?', ['%name' => $this->entity->label()]);
   }
 
   /**
    * {@inheritdoc}
    */
   public function getDescription() {
-    return $this->t('Disabling a mapping will stop any automatic synchronization and hide the mapping.');
+    return $this->t('Enabling a mapping will restart any automatic synchronization.');
   }
 
   /**
@@ -36,7 +36,7 @@ class SalesforceMappingDisableForm extends EntityConfirmFormBase {
    * {@inheritdoc}
    */
   public function getConfirmText() {
-    return $this->t('Disable');
+    return $this->t('Enable');
   }
 
   /**
@@ -45,9 +45,10 @@ class SalesforceMappingDisableForm extends EntityConfirmFormBase {
   public function submit(array $form, FormStateInterface $form_state) {
     parent::submit($form, $form_state);
 
-    $this->entity->disable()->save();
+    $this->entity->enable()->save();
     $form_state['redirect_route'] = [
-      'route_name' => 'entity.salesforce_mapping.list',
+      'route_name' => 'entity.salesforce_mapping.edit_form',
+      'route_parameters' => ['salesforce_mapping' => $this->entity->id()],
     ];
   }
